02-improve-demo-applications/microservices/AWS_lambda_clothoid_path_planner/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Werkzeug==2.2.2
requests==2.28.2
urllib3==1.26.15
pyclothoids==0.1.4
pydantic-core
fastapi
mangum
aws-wsgi